Effects.IORef

Description

Provides the MonadIORef class.

Since: 0.1

Effect

class Monad m => MonadIORef m where

IORef effect.

Since: 0.1

Methods

newIORef :: HasCallStack => a -> m (IORef a)

Build a new IORef.

Since: 0.1

readIORef :: HasCallStack => IORef a -> m a

Read the value of an IORef.

Since: 0.1

writeIORef :: HasCallStack => IORef a -> a -> m ()

Write a new value into an IORef.

Since: 0.1

atomicWriteIORef :: HasCallStack => IORef a -> a -> m ()

Variant of writeIORef with the "barrier to reordering" property that atomicModifyIORef has.

Since: 0.1

modifyIORef' :: HasCallStack => IORef a -> (a -> a) -> m ()

Strict version of modifyIORef.

Since: 0.1

atomicModifyIORef' :: HasCallStack => IORef a -> (a -> (a, b)) -> m b

Strict version of atomicModifyIORef. This forces both the value stored in the IORef and the value returned. The new value is installed in the IORef before the returned value is forced. So

atomicModifyIORef' ref (x -> (x+1, undefined))

will increment the IORef and then throw an exception in the calling thread.

Since: 0.1

Functions

atomicModifyIORef'_ :: MonadIORef m => IORef a -> (a -> a) -> m ()

Atomically apply a function to the contents of an IORef and return the old and new values. The result of the function is forced.

Since: 0.1

Reexports

data IORef a

A mutable variable in the IO monad.

>>> import Data.IORef
>>> r <- newIORef 0
>>> readIORef r
0
>>> writeIORef r 1
>>> readIORef r
1
>>> atomicWriteIORef r 2
>>> readIORef r
2
>>> modifyIORef' r (+ 1)
>>> readIORef r
3
>>> atomicModifyIORef' r (\a -> (a + 1, ()))
>>> readIORef r
4

See also STRef and MVar.